feat: averageAll过滤条件 等级数<3 or 干员数 <10
