+++ src/user-tier-list/user-tier-list.service.ts
@@ -7,8 +7,8 @@ import { CreateUserTierListDTO, EditUserTierListDTO } from './user-tier-list.dto
 import { UserTierList, UserTierListSimple } from './user-tier-list.interface';
 import { formatUserTierList } from 'utils';
 
-//stat
-import { intervalMapping } from 'utils/stat-utils';
+//统计数据相关导入
+import { intervalMapping, isValidTierList } from './stat-utils/stat-average';
 import { Tier, AllTierListStatistics, Statistic, CharStatistics } from './user-tier-list.interface';
 
 
@@ -73,41 +73,26 @@ export class UserTierListService {
     (await this.userTierListModel.find()).map((item) => allTierList.push(item.tierList));
     
     let avgTier : CharStatistics = {};
-    let validCount : number = 0;
+    let validCount : number = 0; //valid的判断方法见stat-utils/stat-average.ts filters
 
     allTierList.forEach((tierList: Tier[]) => {
-      if(tierList.length < 3){ //评级少于3个，不计入统计
-        return;
-      }
+      if(isValidTierList(tierList) == false) return;
 
       validCount += 1;
 
+      /*分数转换*/
       let values :Array<number> = []; //一个用户的所有评级 
       tierList.forEach((tier:Tier)=>{
         values.push(tier.value);
       });
-      
-
-      // this.logger.debug( "Before mapping ", values);
-      
       values = intervalMapping(values, 0, 5); //[min, max] 映射到[0,5]
-      // this.logger.debug( "After mapping: ", values);
       values = values.map(item => 5-item) //reverse，越接近5评分越高
-      // this.logger.debug( "Finally: ", values);
       tierList.forEach((tier, i) =>{
         tier.value = values[i];
       }); //value写回到tier中
 
       tierList.forEach((tier: Tier)=>{
-        /* tier example:
-        {
-          "value": 0,
-          "characterKeys": [
-            "char_263_skadi"
-          ]
-        }
-        */
-        tier.characterKeys.forEach((optId) => {
+        tier.characterKeys.forEach((optId) => { //optId 干员名 e.g.:"char_263_skadi"
           if((optId in avgTier) == false){
             avgTier[optId] = {
               avgValue: 0,
